load enum effect type
