+++ src/states/BattleInitialState.py
@@ -21,7 +21,7 @@ class BattleInitialState(BaseState):
         self.turn = params['turn']
         self.currentTurnOwner = params['currentTurnOwner']  
 
-        # Mock move entities
+        # For setting up the initial position at the start of each battle
         self.player.move_to(self.field[self.player.fieldTile_index], self.field)
         self.enemy.move_to(self.field[self.enemy.fieldTile_index], self.field)
